Fixed a bug in cipher_stealing_encrypt in the computation of the last tweak.
It was using the input iv value, while it should be using the tweak
that was calculated in the step before using calculate_tweak.

Used the test vectors used in aes_xts_decrypt_spec.ml after modifying
the key schedule of the data key to be as should be used in encryption.
The one used in decryption is different so as to accommodate EQINVCIPHER;
a more efficient decryption algorithm defined in FIPS 197 and used in
AWS-LC.